refactor(auth): align password reset and verify-email API error handling
- Extract shared rateLimitUserMessage for forgot and reset flows

- Handle 429 on reset with Retry-After; include status in generic fallback

- Parse verify-email errors via parseApiErrorBody

- Update SETUP_PROGRESS Fase B checklist

Made-with: Cursor

Refs #16
